refactor(reindex): extract reindexTracker into its own file
229 item 15 (partial). inverted_reindex_task_generic.go was 3,423 lines;
Marcin flagged it as a smell. Pull the reindexTracker interface +
fileReindexTracker implementation + its parse helpers (~675 lines) into
inverted_reindex_tracker.go — they form a self-contained abstraction
(on-disk sentinel-based state machine, no dependency on the task code).
Main file drops to 2,744 lines.

Remaining decomposition (LSM hooks, runtime swap, bucket helpers) deferred —
they share more state with ShardReindexTaskGeneric and need more care.
This commit is a pure mechanical move: no symbols renamed, no behavior
changed. Test coverage unchanged (no test files touched).

Refs: weaviate/0-weaviate-issues#229
